refactor: extract resolveTabQueryStringValues to own module
