+++ resources/inertia/Providers/Tabs/TabsProvider.tsx
@@ -1,20 +1,9 @@
 import { useEffect, useRef, useState } from "react";
 import TabsContext from "./TabsContext";
+import { resolveTabQueryStringValues } from "./resolveTabQueryStringValues";
 import { ITab, ITabsContextType, ITabsQueryString, TabChangedMethod } from "./types";
 import Wrapper from "@/Components/Tabs/Wrapper";
 import { router } from "@inertiajs/react";
-
-export function resolveTabQueryStringValues(
-    defaults: ITabsQueryString,
-    tab: string,
-    url: URL,
-): Record<string, string | number | boolean> {
-    const tabDefaults = defaults[tab];
-
-    return Object.fromEntries(
-        Object.entries(tabDefaults).map(([param, value]) => [param, (url.searchParams.get(param) ?? value).toString()]),
-    );
-}
 
 export default function TabsProvider({
     defaultSelected,
