+++ resources/inertia/Providers/Tabs/__tests__/resolveTabQueryStringValues.test.ts
@@ -1,4 +1,4 @@
-import { resolveTabQueryStringValues } from "../TabsProvider";
+import { resolveTabQueryStringValues } from "../resolveTabQueryStringValues";
 import { ITabsQueryString } from "../types";
 
 describe("resolveTabQueryStringValues", () => {
